requirements.txt
django>=1.5.1
django-dingos>=0.1.0
# Additional requirements go here
